requirements.txt
# Librerías necesarias

python-telegram-bot==20.5  # Verifica la última versión compatible
